requirements.txt
Pillow
numpy

